Moved some repeated logic into helper functions
